Move NAudio to Hourglass
