--- Hourglass.NAudio/Hourglass.NAudio.csproj
@@ -1,24 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-    <OutputType>Library</OutputType>
-    <LangVersion>preview</LangVersion>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-  <PropertyGroup>
-    <AssemblyTitle>Hourglass NAudio</AssemblyTitle>
-    <Product>Hourglass NAudio</Product>
-    <Company>Ivan Ivon</Company>
-    <NeutralLanguage>en-US</NeutralLanguage>
-    <Copyright>Copyright © 2024-$([System.DateTime]::Now.Year) Ivan Ivon</Copyright>
-    <AssemblyVersion>$(Version)</AssemblyVersion>
-    <FileVersion>$(Version)</FileVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DisableDiagnosticTracing>true</DisableDiagnosticTracing>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NAudio" Version="2.2.1" />
-    <PackageReference Include="NAudio.Vorbis.Latest" Version="1.5.1" />
-  </ItemGroup>
-</Project>
