+++ Hourglass/Hourglass.csproj
@@ -97,6 +97,7 @@
     <None Include="Resources\Usage.txt" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Hourglass.NAudio\Hourglass.NAudio.csproj" />
+    <PackageReference Include="NAudio" Version="2.2.1" />
+    <PackageReference Include="NAudio.Vorbis.Latest" Version="1.5.1" />
   </ItemGroup>
 </Project>
